PR: Front end update
