+++ src/main/java/com/qa/main/Bae15SpringNsacp1Application.java
@@ -2,6 +2,7 @@ package com.qa.main;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
 @SpringBootApplication
 public class Bae15SpringNsacp1Application {
