--- src/main/resources/static/Axios/account.html
@@ -1,92 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <title>Caniformia CyberPark</title>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/css/bootstrap.min.css">
-  <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.0/dist/jquery.slim.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.1/dist/js/bootstrap.bundle.min.js"></script>
-  <style> 
-    body {
-      background-image: url("./imgs/greentexture.jpg");
-      background-color: #84b693;
-      background-size: cover;
-    }
-    </style>
-</head>
-<header>
-    <nav class="navbar navbar-expand-lg navbar-light bg-light">
-        <a class="navbar-brand" href="#">Navbar</a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-          <span class="navbar-toggler-icon"></span>
-        </button>
-      
-        <div class="collapse navbar-collapse" id="navbarSupportedContent">
-          <ul class="navbar-nav mr-auto">
-            <li class="nav-item active">
-              <a class="nav-link" href="./index.html">Home <span class="sr-only"></span></a>
-            </li>
-            <li class="nav-item">
-                <h5><a class="nav-link" href="https://www.wwf.org.uk/">WWF</a></h5>
-              </li>
-            <li class="nav-item dropdown">
-              <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                Site Quick Nav
-              </a>
-              <div class="dropdown-menu" aria-labelledby="navbarDropdown">
-                <a class="dropdown-item" href="./index.html">Home</a>
-                <a class="dropdown-item" href="./indexF.html">Infosheet</a>
-                <div class="dropdown-divider"></div>
-                <a class="dropdown-item" href="./account.html">Sign up</a>
-              </div>
-            </li>
-            <li class="nav-item">
-              <a class="nav-link disabled" href="#"></a>
-            </li>
-          </ul>
-          <form class="form-inline my-2 my-lg-0">
-            <input class="form-control mr-sm-2" type="search" placeholder="Search" aria-label="Search">
-            <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
-          </form>
-        </div>
-      </nav>
-</header>
-<div class="container-fluid">
-    <h1><p class="text-success"> Create Park Account</p></h1>
-</div>
-<div class="container mt-3">
-    <form action="/action_page.php" class="was-validated">
-        <div class="mb-3 mt-3">
-            <label for="uname" class="form-label">Username:</label>
-            <input type="text" class="form-control" id="uname" placeholder="Enter username" name="uname" required>
-            <div class="valid-feedback">Valid.</div>
-            <div class="invalid-feedback">Please fill out this field.</div>
-        </div>
-        <div class="mb-3">
-            <label for="pwd">Password:</label>
-            <input type="password" class="form-control" id="pwd" placeholder="Enter password" name="pswd" required>
-            <div class="valid-feedback">Valid.</div>
-            <div class="invalid-feedback">Please fill out this field.</div>
-        </div>
-        <div class="mb-3">
-            <label for="pwd">Email Address:</label>
-            <input type="email" class="form-control" id="email" placeholder="Enter email" name="email" required>
-            <div class="valid-feedback">Valid.</div>
-            <div class="invalid-feedback">Please fill out this field.</div>
-        </div>
-        <div class="form-check mb-3">
-            <h3><p class="text-primary"><label>Gender:</label></h3>
-            <h4><input type="radio" id="m_gender" value="m_gender" name="gender"><label for="m_gender"> Male </label></h4>
-            <h4><input type="radio" id="f_gender" value="f_gender" name="gender"><label for="f_gender"> Female </label></h4>
-            <div>
-                <label class="form-check-label">
-                    <input class="form-check-input" type="checkbox" name="remember"> Remember me
-                </label>
-            </div>
-        </div>
-        <button type="Submit" class="btn btn-primary">Submit Info</button>
-    </form>
-</body>
-</html>
