--- src/main/resources/static/Axios/css/style.css
@@ -1,44 +1,0 @@
-#formDiv {
-    margin-top: 3%;
-    margin-left: 2%;
-    width: 25%;
-    border: solid 2px rgb(131, 84, 46);
-    display: inline-grid;
-    padding: 10px;
-}
-
-#results-div {
-    width: 50%;
-    height:fit-content;
-    display: inline-grid;
-    margin-top: 3%;
-    margin-left: 5%;
-    padding: 10px;
-    border:hidden;
-    background-color:transparent;
-}   
-
-.entry-parent {
-    padding-left: 30%;
-}
-
-.entry-div {
-    border-style: solid;
-    border-width: 2px;
-    border-radius: 6px;
-    height: 2em;
-    background-color: rgb(30, 97, 75);
-    border-color: rgb(18, 48, 48);
-    color: azure;
-    margin-top: 10px;
-    width: 100%;
-    text-align: center;
-    display: inline-grid;
-}
-
-#idInput {
-    display: inline-grid;
-    margin-top: 3%;
-    margin-left: 2%;
-}
-
